FEAT: all projectile types without damage logic
