+++ Assets/test-devgame/Prefabs/Projectiles/GreandeLauncherAmmo.prefab
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1001 &5017681653924324197
+--- !u!1001 &5341833313723893720
 PrefabInstance:
   m_ObjectHideFlags: 0
   serializedVersion: 2
@@ -8,13 +8,21 @@ PrefabInstance:
     serializedVersion: 3
     m_TransformParent: {fileID: 0}
     m_Modifications:
-    - target: {fileID: 3418119373784925532, guid: 8424af50deaefb24d9a3d526d5e3db71, type: 3}
-      propertyPath: m_Name
-      value: ShotgunAmmo
+    - target: {fileID: -8596299237265507846, guid: 8424af50deaefb24d9a3d526d5e3db71, type: 3}
+      propertyPath: damage
+      value: 10
+      objectReference: {fileID: 0}
+    - target: {fileID: -8596299237265507846, guid: 8424af50deaefb24d9a3d526d5e3db71, type: 3}
+      propertyPath: maxRange
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -8596299237265507846, guid: 8424af50deaefb24d9a3d526d5e3db71, type: 3}
+      propertyPath: explosiveRange
+      value: 2
       objectReference: {fileID: 0}
     - target: {fileID: 3418119373784925532, guid: 8424af50deaefb24d9a3d526d5e3db71, type: 3}
-      propertyPath: m_Layer
-      value: 3
+      propertyPath: m_Name
+      value: GreandeLauncherAmmo
       objectReference: {fileID: 0}
     - target: {fileID: 7997782438656472627, guid: 8424af50deaefb24d9a3d526d5e3db71, type: 3}
       propertyPath: m_LocalPosition.x
@@ -56,32 +64,8 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
-    m_RemovedComponents:
-    - {fileID: -8596299237265507846, guid: 8424af50deaefb24d9a3d526d5e3db71, type: 3}
+    m_RemovedComponents: []
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
-    m_AddedComponents:
-    - targetCorrespondingSourceObject: {fileID: 3418119373784925532, guid: 8424af50deaefb24d9a3d526d5e3db71, type: 3}
-      insertIndex: 2
-      addedObject: {fileID: -1657313450060323442}
+    m_AddedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 8424af50deaefb24d9a3d526d5e3db71, type: 3}
---- !u!1 &7696083954548581945 stripped
-GameObject:
-  m_CorrespondingSourceObject: {fileID: 3418119373784925532, guid: 8424af50deaefb24d9a3d526d5e3db71, type: 3}
-  m_PrefabInstance: {fileID: 5017681653924324197}
-  m_PrefabAsset: {fileID: 0}
---- !u!114 &-1657313450060323442
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7696083954548581945}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: c769718956cae134eb3bec77e97836b7, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  speed: 4
-  damage: 2
-  maxRange: 7
