+++ Assets/test-devgame/Scripts/Abstract/Projectile.cs
@@ -1,9 +1,33 @@
+using System.Collections;
 using UnityEngine;
 
 public abstract class Projectile : MonoBehaviour
 {
     [SerializeField] protected float speed;
     [SerializeField] private float damage;
+    
+    [SerializeField] private float maxRange;
+    
+    private void Start()
+    {
+        if (maxRange > 0)
+        {
+            float flightTime = maxRange / speed;
+            Destroy(gameObject, flightTime);
+        }
+    }
+
+    private IEnumerator DestroyOnTargetReached(float flightTime)
+    {
+        yield return new WaitForSeconds(flightTime);
+        OnTargetReached();
+        Destroy(gameObject);
+    }
+
+    protected virtual void OnTargetReached()
+    {
+        
+    }
     
     protected virtual void Update()
     {
@@ -15,5 +39,4 @@ public abstract class Projectile : MonoBehaviour
     {
         transform.Translate(Vector2.right * (speed * Time.deltaTime));
     }
-    protected abstract void OnCollisionEnter2D(Collision2D collision);
 }
