+++ Assets/test-devgame/Scripts/Projectiles/DefaultProjectile.cs
@@ -1,0 +1,10 @@
+using UnityEngine;
+
+public class DefaultProjectile : Projectile
+{
+    // TODO: add damage logic
+    protected void OnCollisionEnter2D(Collision2D collision)
+    {
+        Destroy(gameObject);
+    }
+}
