+++ Assets/test-devgame/Scripts/Projectiles/ExplosiveProjectile.cs
@@ -1,0 +1,17 @@
+using UnityEngine;
+
+public class ExplosiveProjectile : Projectile
+{
+    [SerializeField] private float explosiveRange;
+    
+    protected override void OnTargetReached()
+    {
+        Explode();
+    }
+
+    // TODO: add explosive damage logic
+    private void Explode()
+    {
+        
+    }
+}
