--- Assets/test-devgame/Scripts/Projectiles/UnlimitedRangeBullet.cs
@@ -1,9 +1,0 @@
-using UnityEngine;
-
-public class UnlimitedRangeBullet : Projectile
-{
-    protected override void OnCollisionEnter2D(Collision2D collision)
-    {
-        Destroy(gameObject);
-    }
-}
